Conflicts:
	kexi/main/KexiMainWindow_p.h

+++ krita/plugins/filters/sobelfilter/kis_sobel_filter.cpp
@@ -97,7 +97,7 @@ void KisSobelFilter::prepareRow(KisPaintDeviceSP src, quint8* data, quint32 x, q
     }
 }
 
-#define RMS(a, b) (sqrt ((a) * (a) + (b) * (b)))
+#define RMS(a, b) (sqrt ((qreal)(a) * (a) + (b) * (b)))
 #define ROUND(x) ((int) ((x) + 0.5))
 
 void KisSobelFilter::process(KisPaintDeviceSP device,
